delete insert-specific stuff
