+++ airbyte-integrations/connectors/destination-redshift/src/test-integration/java/io/airbyte/integrations/destination/redshift/RedshiftConnectionTest.java
@@ -17,8 +17,8 @@ import org.junit.jupiter.api.Test;
 
 public class RedshiftConnectionTest {
 
-  private final JsonNode config = Jsons.deserialize(IOs.readFile(Path.of("secrets/config.json")));
-  private final RedshiftDestination destination = new RedshiftDestination();
+  private final JsonNode config = Jsons.deserialize(IOs.readFile(Path.of("secrets/config_staging.json")));
+  private final RedshiftStagingS3Destination destination = new RedshiftStagingS3Destination();
   private AirbyteConnectionStatus status;
 
   @Test
